should be done with /courses and /courses/course_id
